Fixed a bug that caused fire to indirectly damage teammates.
To clarify what was going on here: commit fbb01880 allowed napalm-style fire to afflict any entity with .takedamage set to DAMAGE_AIM, not just players exclusively. One instance of such a type of entity happens to be the MEODP buff "death point"... When a player steps into a "friendly fire", it does no damage to them (unless the actual friendly fire is enabled) as usual -- however, their "death point" is not attached to a team, and so it caches the fire. It not being solid at this point doesn't save it either, because the trace check inside HG_Napalm_Burn only carese about trace_fraction. You'd think that would kill the player instantly, as if the point was shot at by a MEODP user, but no -- Buffs_MEODP_Point_Damage explicitly refuses to do anything if the point is not in solid state (which it is when a meodp user fires a hitscan gun). However, fire damage spreads to nearby entities in this game, and whatever is burning is deemed the "owner" of the damage it spreads, regardless of what or who has set it aflame. Guess what usually happens to be the closest entity to a "burning" death point belonging to a player, eh?

The fix is fairly simple - have the death points take damage only when they are solid. This seems like the right thing to do, anyway.