Added "tracking waypoints": a special kind of personal waypoints that are attached to teammates.
Usage:

    cmd waypoint track #id [text] [color]

id is the player id of the teammate you want to track, as seen in status.
You can also supply "#target" instead to use the player you're currently looking at (useful for binds).

text is the text to display above the player.
You can leave it blank to use the player's name (decolorized).

color is an RGB vector, values in range [0..1].
e.g. "1 0 0" for a red waypoint.
The default is white ("1 1 1").

If a tracking waypoint for the desired teammate is already set, it is replaced, but only if the text parameter is supplied and isn't empty; otherwise the waypoint removed.