Made use of the DP_CSQC_QUERYRENDERENTITY extension to exactly attach overhead buff models to players that own them (this way they don't lag behind).
Clients that don't support this feature will fall back to the old method.

In addition, this lets the buff models benifit from playermodel culling, potentially increasing performance.
If the player model is culled (definitely invisible), the buff model won't be rendered, but the dynamic light from it will still be there, because it might be possible to see it if the player is just around a corner.
If the rm_fps_friendly is enabled, then the dynamic light will be culled together with the buff model.

TODO: the same thing for ent_cs stuff (e.g. shownames)