Added an "UI scale" setting - an easier way to customize vid_conwidth and vid_conheight (virtual resolution for 2D stuff, e.g. the HUD), without having to worry about the aspect ratio.
You can set it in the menu, Settings -> Video. Or use the cvar, vid_ui_scale.
Note that a recent version of DPRM is required to apply changes immediately, and the RM menu must be loaded for the cvar to have any effect.

The default value is 1, which matches the standard Nexuiz behaviour exactly.
For high resolutions, a slightly lower value is recommended to save some screen space, try 0.75.
Setting it over 1 is possible but not recommended, because it can make the HUD no longer fit on the screen.
Disabling UI scaling will cause the 2D resolution to match the real resolution exactly.